Angular Commit

+++ src/app/Views/index-contrat/index-contrat.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ContratModule } from 'src/app/Models/contrat/contrat.module';
 import { ContratService } from 'src/app/Service/contrat.service';
+import { ConfirmationDialogService } from 'src/app/Views/confirmation-dialog/confirmation-dialog.service';
 
 @Component({
   selector: 'app-index-contrat',
@@ -9,7 +10,7 @@ import { ContratService } from 'src/app/Service/contrat.service';
 })
 export class IndexContratComponent implements OnInit {
   contrats: ContratModule[];
-  constructor(private contratService: ContratService) { }
+  constructor(private contratService: ContratService , private confirmationDialogService: ConfirmationDialogService) { }
 
   ngOnInit(): void {
     this.getAllContrat();
@@ -25,8 +26,17 @@ export class IndexContratComponent implements OnInit {
           console.log(error);
         });
   }
-  delete(id: number) {
-    this.contratService.deleteContrat(id);
-    this.getAllContrat();
-   }
+   public openConfirmationDialog(id: number) {
+    this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
+    .then((confirmed) => {
+     if (confirmed === true){this.contratService.deleteContrat(id).subscribe(
+      response => {
+        console.log(response);
+        this.getAllContrat();
+      },
+      error => {
+        console.log(error);
+      }); }
+    }).catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+  }
 }
