+++ src/app/Views/index-maison/index-maison.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { MaisonModule } from 'src/app/Models/maison/maison.module';
 import { MaisonService } from 'src/app/Service/maison.service';
+import { ConfirmationDialogService } from 'src/app/Views/confirmation-dialog/confirmation-dialog.service';
 
 
 @Component({
@@ -10,7 +11,7 @@ import { MaisonService } from 'src/app/Service/maison.service';
 })
 export class IndexMaisonComponent implements OnInit {
   maisons: MaisonModule[];
-  constructor(private maisonService: MaisonService) { }
+  constructor(private maisonService: MaisonService , private confirmationDialogService: ConfirmationDialogService) { }
 
   ngOnInit(): void {
     this.getAllMaison();
@@ -26,13 +27,17 @@ export class IndexMaisonComponent implements OnInit {
           console.log(error);
         });
   }
-  delete(id: number) {
-    this.maisonService.deleteMaison(id).subscribe(
+   public openConfirmationDialog(id: number) {
+    this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
+    .then((confirmed) => {
+     if (confirmed === true){this.maisonService.deleteMaison(id).subscribe(
       response => {
         console.log(response);
+        this.getAllMaison();
       },
       error => {
         console.log(error);
-      });
-   }
+      }); }
+    }).catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+  }
 }
