+++ src/app/Views/index-marque/index-marque.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { MarqueModule } from 'src/app/Models/marque/marque.module';
 import { MarqueService } from 'src/app/Service/marque.service';
+import { ConfirmationDialogService } from 'src/app/Views/confirmation-dialog/confirmation-dialog.service';
 
 @Component({
   selector: 'app-index-marque',
@@ -9,7 +10,7 @@ import { MarqueService } from 'src/app/Service/marque.service';
 })
 export class IndexMarqueComponent implements OnInit {
   marques: MarqueModule[];
-  constructor(private marqueService: MarqueService) { }
+  constructor(private marqueService: MarqueService , private confirmationDialogService: ConfirmationDialogService) { }
 
   ngOnInit(): void {
     this.getAllMarque();
@@ -25,8 +26,17 @@ export class IndexMarqueComponent implements OnInit {
           console.log(error);
         });
   }
-  delete(id: number) {
-    this.marqueService.deleteMarque(id);
-    this.getAllMarque();
-   }
+   public openConfirmationDialog(id: number) {
+    this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
+    .then((confirmed) => {
+     if (confirmed === true){this.marqueService.deleteMarque(id).subscribe(
+      response => {
+        console.log(response);
+        this.getAllMarque();
+      },
+      error => {
+        console.log(error);
+      }); }
+    }).catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+  }
 }
