+++ src/app/Views/index-societe/index-societe.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { SocieteModule } from 'src/app/Models/societe/societe.module';
 import { SocieteService } from 'src/app/Service/societe.service';
+import { ConfirmationDialogService } from 'src/app/Views/confirmation-dialog/confirmation-dialog.service';
 
 @Component({
   selector: 'app-index-societe',
@@ -9,7 +10,7 @@ import { SocieteService } from 'src/app/Service/societe.service';
 })
 export class IndexSocieteComponent implements OnInit {
   societes: SocieteModule[];
-  constructor(private societeService: SocieteService) { }
+  constructor(private societeService: SocieteService , private confirmationDialogService: ConfirmationDialogService) { }
 
   ngOnInit(): void {
     this.getAllSociete();
@@ -25,8 +26,17 @@ export class IndexSocieteComponent implements OnInit {
           console.log(error);
         });
   }
-  delete(id: number) {
-    this.societeService.deleteSociete(id);
-    this.getAllSociete();
-   }
+   public openConfirmationDialog(id: number) {
+    this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
+    .then((confirmed) => {
+     if (confirmed === true){this.societeService.deleteSociete(id).subscribe(
+      response => {
+        console.log(response);
+        this.getAllSociete();
+      },
+      error => {
+        console.log(error);
+      }); }
+    }).catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+  }
 }
